requirements.txt
streamlit
google-generativeai
python-dotenv
pypdf
requests
fastapi
uvicorn
python-multipart
